requirements.txt
fastapi
uvicorn
pandas
joblib
python-dotenv
pymongo
scikit-learn
xgboost
